requirements.txt
# PSLE Science AI Tutor - Dependencies
# Generated from code imports for Streamlit Community Cloud deployment

# Core Framework
streamlit>=1.28.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0

# Visualization
plotly>=6.0.0

# Image Processing
pillow>=10.0.0

# Environment Variables (for local development)
python-dotenv>=1.0.0

# PDF Support (optional, used in upload script)
PyPDF2>=3.0.0
